Amélioration sécurité check ID

+++ backend/controllers/sauce.js
@@ -1,6 +1,7 @@
 const Sauce = require("../models/Sauce");
 const fs = require("fs");
 const { parse } = require("ts-node");
+const { JsonWebTokenError } = require("jsonwebtoken");
 
 exports.createSauce = (req, res, next) => {
     const sauceObject = JSON.parse(req.body.sauce);
@@ -21,23 +22,23 @@ exports.createSauce = (req, res, next) => {
 };
 
 exports.modifySauce = (req, res, next) => {
-    const sauceObject = req.file
-        ? {
-              ...JSON.parse(req.body.sauce),
-              imageUrl: `${req.protocol}://${req.get("host")}/images/${req.file.filename}`,
-          }
-        : {
-              ...req.body,
-          };
+    if (req.file) {
+        const sauceObject = { ...JSON.parse(req.body.sauce), imageUrl: `${req.protocol}://${req.get("host")}/images/${req.file.filename}` };
 
-    Sauce.findOne({ _id: req.params.id }).then((sauce) => {
-        const filename = sauce.imageUrl.split("/images/")[1];
-        fs.unlink(`images/${filename}`, () => {
-            Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
-                .then(() => res.status(200).json({ message: "Sauce modifiée !" }))
-                .catch((error) => res.status(400).json({ error }));
+        Sauce.findOne({ _id: req.params.id }).then((sauce) => {
+            const filename = sauce.imageUrl.split("/images/")[1];
+            fs.unlink(`images/${filename}`, () => {
+                Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+                    .then(() => res.status(200).json({ message: "Sauce modifiée !" }))
+                    .catch((error) => res.status(400).json({ error }));
+            });
         });
-    });
+    } else {
+        const sauceObject = { ...req.body };
+        Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+            .then(() => res.status(200).json({ message: "Sauce modifiée !" }))
+            .catch((error) => res.status(400).json({ error }));
+    }
 };
 
 exports.deleteSauce = (req, res, next) => {
@@ -66,7 +67,6 @@ exports.getAllSauces = (req, res, next) => {
 };
 
 exports.changeLikes = (req, res, next) => {
-    const sauceLiked = req.body;
     const user = req.body.userId;
     const like = req.body.like;
